requirements.txt
streamlit
openai
langchain
faiss-cpu
chromadb
tiktoken
python-dotenv
ragas